Remove duplicate code. Move BackupTLSConfig.* from fdbbackup to fdbclient so can be used in fdbclient. Remove the copies of BackupTLSConfig we had in place named BlobTLSConfig.*. Keep the old name though it a little clunky.
