+++ fdbbackup/CMakeLists.txt
@@ -1,5 +1,4 @@
 set(FDBBACKUP_SRCS
-	BackupTLSConfig.cpp
 	Decode.cpp
 	backup.actor.cpp)
 
@@ -15,7 +14,6 @@ target_link_libraries(fdbconvert PRIVATE fdbclient)
 
 set(FDBDECODE_SRCS
 	Decode.cpp
-	BackupTLSConfig.cpp
 	FileDecoder.actor.cpp)
 add_flow_target(EXECUTABLE NAME fdbdecode SRCS ${FDBDECODE_SRCS})
 target_include_directories(fdbdecode PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/include" "${CMAKE_CURRENT_BINARY_DIR}/include")
